Replace Decimal with float in various fields
Updated all occurrences of Decimal to float for specific fields across multiple files. This includes fields like custom_rate_percent, adjust_percentage, quantity, and rate_percent. Additionally, introduced several new classes and imports to handle updated functionality and new references.

+++ src/quickbooks_desktop/lists/billing_rate.py
@@ -46,14 +46,14 @@ class BillingRatePerItem(QBMixin):
             "type": "Element",
         },
     )
-    custom_rate_percent: Optional[Decimal] = field(
+    custom_rate_percent: Optional[float] = field(
         default=None,
         metadata={
             "name": "CustomRatePercent",
             "type": "Element",
         },
     )
-    adjust_percentage: Optional[Decimal] = field(
+    adjust_percentage: Optional[float] = field(
         default=None,
         metadata={
             "name": "AdjustPercentage",
